# API_documentation.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: COURSE API
  description: This is the API Documentation of Course Recommender System of Team Sept-07
servers:
  - url: http://127.0.0.1:5001
...
paths:
...
  /api/review:
    post:
      tags:
        - Review
      summary: To create new review
      description: Operation to add review.
      requestBody:
        description: Add Review
        content:
          application/json:
            schema:
              type: object
              properties:
                course_id:
                  type: integer
                  example: 1
                user_id:
                  type: integer
                  example: 1
                difficulty:
                  type: integer
                  example: 3
                review:
                  type: string
                  example: bla bla bla
                support:
                  type: integer
                  example: 2
                rating:
                  type: integer
                  example: 2
      responses:
        '201':
          description: Success
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Review Added Successfully"}'
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: >-
                    {"message": "Can't give more than one review for each
                    course"} course"}
                Example 2:
                  value: '{"message": "User has not enrolled for this course"}'
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid course_id"}'
        '500':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something went wrong"}'
  /api/review/{review_id}:
    get:
      tags:
        - Review
      summary: To get review data
      description: Operation to get Review details.
      parameters:
        - in: path
          name: review_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    data:
                      course_id: 1
                      course_name: Java
                      difficulty: 5
                      rating: 4
                      review: blab blabalba
                      review_id: 2
                      support: 3
                      user_id: 5
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid review_id"}'
    put:
      tags:
        - Review
      summary: To updating review data
      description: Operation to update Review Details.
      parameters:
        - in: path
          name: review_id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                review_id:
                  type: integer
                  example: 2
                difficulty:
                  type: integer
                  example: 4
                support:
                  type: integer
                  example: 3
                rating:
                  type: integer
                  example: 2
                review:
                  type: string
                  example: Wonderful
      responses:
        '202':
          description: Success
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Review Successfully Updated"}'
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid Review Id"}'
        '500':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something went wrong"}'
    delete:
      tags:
        - Review
      summary: To delete review
      description: Operation to delete a Review.
      parameters:
        - in: path
          name: review_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Review Deleted Successfully"}'
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid Review Id"}'
        '500':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something went wrong"}      '
  /api/reviews/student/{user_id}:
    get:
      tags:
        - Review
      summary: To get reviews given by student
      description: Operation to get Student's Reviews details.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    data:
                      - course_id: 2
                        course_name: System Commands
                        difficulty: 5
                        rating: 4
                        review: blab blabalba
                        review_id: 3
                        support: 3
                        user_id: 1
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid user_Id"}'
...